Yosemite build alternative fix.
This commit should fix the build for Yosemite OS, which was broken
due to the introduction of conflicting macros.

PR #77 fix the build.
This should also fix the build, but in a less intrusive way.

It prevents the definition of the htonll and ntohll function
inside the zmqpp namespace if a equivalent macro is present.
message.cpp has been update so that it will either call the macro
or the zmqpp function.

However, it is yet to be tested on Yosemite.