CMake can depend on libzmq cmake.
Add a CMake flag (ZMQPP_LIBZMQ_CMAKE). When set to true (which is
not the default value) this flag instructs CMake to link against
libzmq built through CMake: cmake will now properly know that
zmqpp depends and libzmq and will not fail if it can't find it
in the path.