Fix build in some rare case.
Problem: My previous update to CMake was incomplete. It looks
like if zmq.h was present in /usr/local/include it had precedence
over the one from libzmq folder (that we built with CMake).
This patch to CMakeFile make sure that we use the custom include
path first, thus avoiding this problem.

Problem: curve.hpp missed an include of zmq.h, therefore causing
ZMQ_VERSION_MAJOR to be undefined.