zmqpp::auth will now set the user-id in PLAIN and CURVE mode.
When a ZAP request of type PLAIN or CURVE is handled, the ZAP
handler's reply now include a User-Id property.

For CURVE mode, this is the z85 encoded public key of the client.
For PLAIN, the username is used as the User-Id.

The tests of the auth code have been modified to check the new
behavior.