poller::add() takes a const reference now.
The add() method taking a zmq_pollitem_t got its parameter as a value.
The other methods take zmq_pollitem_t as a const reference. In order to
be consistent add() does that too now. Using a const reference is also
more appropriate to avoid an unnecessary copy of the struct.