Fix error in test.
This fixes #116.
Problem: The creation of the zmq_pollitem_t object
had a typo, resulting in a call to zmq_poll() that didn't
really make sense.
Solution: Correctly create the pollitem and adjust the BOOT_CHECK
call to a proper expected result. Also increase the max timeout
to 1second to make sure the message as plenty of time to be
delivered.