Problem: No Copy behavior could be improved.
See #123. The NoCopy mechanism was available through `move()`
and `add_const()`.

This patch adds `add_nocopy()` and `add_nocopy_const()`.
`add_nocopy()` is similar to `move()` but does not perform
memory allocation to create a type safe deleter. Instead the
user provide his own libzmq-level deleter. Reject pointer to
const data.

`add_nocopy_const()` let the user pass a pointer to const data
and perform the cast internally. This function is to make it
very clear that the user pass const data and that he must
therefore be careful not modifying them later.

Deprecate `add_const()` in favor of the new functions (see #123
for more info).